+++ CHANGELOG.md
@@ -12,10 +12,12 @@ at anytime.
   * Prevent publish of files with size 0
   * Add `dht_status` parameter to `status` to include bandwidth and peer info
   *
+  * Positional and flag arguments in lbrynet-cli
+  * 
 
 ### Changed
-  *
-  *
+  * Changed keyword arguments in lbrynet-cli to use a -- prefix
+  * Using the help function in lbrynet-cli no longer requires lbrynet-daemon to be running
 
 ### Fixed
   *